uevent.c: Set nl_pid to 0
If any other uevent sockets are opened before uevent_init, bind will
fail because of a non-unique id. Instead, set the nl_pid to 0.

From man 7 netlink:

If the application sets nl_pid before calling bind(2), then it  is  up
to the application to make sure that nl_pid is unique.  If the
application sets it to 0, the kernel takes care of assigning it.  The
kernel assigns the process ID to the first netlink socket the process
opens and assigns a unique nl_pid to every netlink socket that the
process subsequently creates.

Bug: 434740485
Flag: EXEMPT bugfix
Test: Opened uevent socket twice in system server successfully.
Change-Id: Id725066028ebbc7654e4b13d92ba961d2dc012c1